Start adding playground execution functionality
This is a very WIP commit as I immediately ran into issues with CORS
trying to get requests to fire in dev. Solution is probably to bring
this on domain in dev maybe, or otherwise have a backend proxy for dev
that this can talk to (which might even solve routing issues).